requirements.txt
Django==2.2.7
django-cors-headers==3.2.1
django-filter==2.2.0
djangorestframework==3.10.3
djangorestframework-simplejwt==4.4.0
numpy==1.18.1
Pillow==6.2.1
psycopg2==2.8.4
PyJWT==1.7.1
pytz==2019.3
sqlparse==0.3.0
xhtml2pdf==0.2.4
num2words==0.5.10